refact: PhotoSaveService 함수 뷰모델로 이동 후 파일 삭제
